feat(Google AI Gemini) — applying patch from Dmytro
