+++ langchain4j/src/test/java/dev/langchain4j/service/OpenAiAiServicesWithJsonSchemaIT.java
@@ -1,0 +1,41 @@
+package dev.langchain4j.service;
+
+import dev.langchain4j.model.chat.ChatLanguageModel;
+import dev.langchain4j.model.openai.OpenAiChatModel;
+
+import java.util.List;
+
+import static dev.langchain4j.model.openai.OpenAiChatModelName.GPT_4_O_MINI;
+import static java.util.Arrays.asList;
+
+// TODO move to langchain4j-open-ai module once dependency cycle is resolved
+class OpenAiAiServicesWithJsonSchemaIT extends AiServicesWithJsonSchemaIT {
+
+    @Override
+    protected List<ChatLanguageModel> models() {
+        return asList(
+                OpenAiChatModel.builder()
+                        .baseUrl(System.getenv("OPENAI_BASE_URL"))
+                        .apiKey(System.getenv("OPENAI_API_KEY"))
+                        .organizationId(System.getenv("OPENAI_ORGANIZATION_ID"))
+                        .modelName(GPT_4_O_MINI)
+                        .responseFormat("json_schema")
+                        .strictJsonSchema(true)
+                        .temperature(0.0)
+                        .logRequests(true)
+                        .logResponses(true)
+                        .build(),
+                OpenAiChatModel.builder()
+                        .baseUrl(System.getenv("OPENAI_BASE_URL"))
+                        .apiKey(System.getenv("OPENAI_API_KEY"))
+                        .organizationId(System.getenv("OPENAI_ORGANIZATION_ID"))
+                        .modelName(GPT_4_O_MINI)
+                        .responseFormat("json_schema")
+                        .strictJsonSchema(false)
+                        .temperature(0.0)
+                        .logRequests(true)
+                        .logResponses(true)
+                        .build()
+        );
+    }
+}
